modify: move `models.py` into `dev` folder
