+++ apps/draw/draw_funcs/profile.py
@@ -8,7 +8,7 @@ from PIL import Image, ImageChops, ImageDraw
 import apps.draw.utility as draw_utility
 import asset
 from apps.text_map import text_map
-from models import DynamicBackgroundInput, TopPadding
+from dev.models import DynamicBackgroundInput, TopPadding
 
 
 def character_card(
