+++ apps/genshin/enka.py
@@ -6,7 +6,7 @@ import asyncpg
 import enkanetwork
 
 from exceptions import NoCharacterFound
-from models import EnkaInfoResponse
+from dev.models import EnkaInfoResponse
 
 
 async def get_enka_info(uid: int, session: aiohttp.ClientSession) -> EnkaInfoResponse:
