+++ apps/genshin/utility.py
@@ -13,7 +13,7 @@ from discord import Locale
 from discord.utils import format_dt, get
 
 import asset
-import models
+import dev.models as models
 from ambr import AmbrTopAPI, Character, Domain, Material, Weapon
 from ambr.models import CharacterDetail
 from apps.db import get_user_lang
