+++ base_ui.py
@@ -10,7 +10,7 @@ import asset
 import exceptions
 from apps.db import get_user_lang
 from apps.text_map import text_map
-from models import CustomInteraction, ErrorEmbed, OriginalInfo
+from dev.models import CustomInteraction, ErrorEmbed, OriginalInfo
 from utility import log
