+++ cogs/schedule.py
@@ -13,7 +13,7 @@ from discord.ext import commands, tasks
 import ambr
 import apps.genshin as genshin_app
 import asset
-import models
+import dev.models as models
 from apps.db import get_user_lang, get_user_theme
 from apps.draw import main_funcs
 from apps.text_map import text_map, to_ambr_top
