+++ ui/genshin/Abyss.py
@@ -9,7 +9,7 @@ from apps.draw import main_funcs
 from apps.draw.main_funcs import draw_abyss_one_page
 from apps.text_map import text_map
 from base_ui import BaseView
-from models import AbyssResult, CustomInteraction, DefaultEmbed, DrawInput
+from dev.models import AbyssResult, CustomInteraction, DefaultEmbed, DrawInput
 
 
 class View(BaseView):
