+++ ui/genshin/Domain.py
@@ -13,7 +13,7 @@ from apps.draw import main_funcs
 from apps.genshin import get_farm_data, get_uid, get_uid_tz
 from apps.text_map import text_map
 from base_ui import BaseView
-from models import DefaultEmbed, DrawInput
+from dev.models import DefaultEmbed, DrawInput
 from utility import get_dt_now
