+++ ui/genshin/MeToo.py
@@ -5,7 +5,7 @@ from apps.db.utility import get_user_lang
 from apps.genshin import check_cookie_predicate
 from apps.text_map import text_map
 from base_ui import BaseView
-from models import CustomInteraction
+from dev.models import CustomInteraction
 
 
 class View(BaseView):
