+++ ui/genshin/ReminderMenu.py
@@ -14,7 +14,7 @@ from apps.genshin import (
 from apps.text_map import text_map, to_ambr_top
 from base_ui import BaseModal, BaseView
 from exceptions import NumbersOnly
-from models import DefaultEmbed
+from dev.models import DefaultEmbed
 from ui.genshin import TalentNotificationMenu, WeaponNotificationMenu
 from utility import divide_chunks
