+++ ui/others/ManageAccounts.py
@@ -14,7 +14,7 @@ from apps.genshin import get_account_select_options, get_uid_region_hash
 from apps.text_map import text_map, to_hutao_login_lang
 from base_ui import BaseModal, BaseView
 from cogs.login import register_user
-from models import CustomInteraction, DefaultEmbed, ShenheBot
+from dev.models import CustomInteraction, DefaultEmbed, ShenheBot
 from utility import log
