+++ ui/others/settings/CustomImage.py
@@ -11,7 +11,7 @@ from apps.genshin import get_character_emoji
 from apps.text_map import text_map, to_ambr_top
 from base_ui import BaseModal, BaseView, EnkaView
 from data.game.elements import get_element_emoji, get_element_list
-from models import CustomInteraction, DefaultEmbed, ErrorEmbed
+from dev.models import CustomInteraction, DefaultEmbed, ErrorEmbed
 from ui.genshin import EnkaDamageCalc
 from utility import divide_chunks
