+++ ui/others/settings/Notif.py
@@ -6,7 +6,7 @@ import config
 from apps.db import get_user_notif
 from apps.text_map import text_map
 from base_ui import BaseView, GoBackButton
-from models import CustomInteraction, DefaultEmbed, OriginalInfo
+from dev.models import CustomInteraction, DefaultEmbed, OriginalInfo
 
 
 class View(BaseView):
