+++ ui/todo/TodoList.py
@@ -6,7 +6,7 @@ from discord.ui import Button, Select, TextInput
 
 import asset
 import config
-import models
+import dev.models as models
 from ambr import AmbrTopAPI, Material
 from apps.db import get_user_lang, get_user_theme
 from apps.draw import main_funcs
