+++ ui/wish/ChooseBanner.py
@@ -3,11 +3,11 @@ from typing import List
 import discord
 
 import config
-import models
+import dev.models as models
 from apps.db import get_user_lang, get_user_theme
 from apps.draw import main_funcs
 from base_ui import BaseView
-from models import DefaultEmbed, DrawInput
+from dev.models import DefaultEmbed, DrawInput
 
 
 class View(BaseView):
