+++ utility/todo_paginator.py
@@ -6,7 +6,7 @@ from discord.ui import Button, Select
 
 from ambr import Material
 from apps.draw import main_funcs
-from models import CustomInteraction, DrawInput, TodoItem
+from dev.models import CustomInteraction, DrawInput, TodoItem
 from utility.paginator import GeneralPaginator, GeneralPaginatorView
